Afegida table de generadors a gameView i implementada logica perque s'actualitzi

+++ src/Presentation/Controllers/ThreadController.java
@@ -1,5 +1,5 @@
 package Presentation.Controllers;
 
 public interface ThreadController {
-    void updateStoreCurrency(int amount);
+    void updateStoreCurrency(float amount);
 }
